+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfigurationSections/MetricsConfiguration.tsx
