+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfigurationSections/SortConfiguration.tsx
