+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfigurationSections/configurationFields/InputField.tsx
@@ -18,7 +18,7 @@ import * as React from 'react';
 import { useCallback } from 'react';
 
 import { Input } from 'components/bootstrap';
-import { FieldComponentProps } from 'views/components/aggregationwizard/elementConfiguration/VisualizationConfigurationOptions';
+import { FieldComponentProps } from 'views/components/aggregationwizard/elementConfigurationSections/VisualizationConfigurationOptions';
 
 type Props = FieldComponentProps & {
   type: 'numeric',
