+++ graylog2-web-interface/src/views/components/aggregationwizard/elementConfigurationSections/configurationFields/SelectField.tsx
